split test into different classes, update test code for contact management
